add new controller based on workload controller
